utils/requirements.txt
textual>=0.44.0
rich>=13.0.0
pyyaml>=6.0
requests>=2.28.0
pexpect>=4.9.0
pytest>=7.4.0
pytest-mock>=3.12.0
